ObjectSearchServiceAsync: Compiler-fehler behoben, ResultInfoBox: keine Anzeige einer neg. Gesamtanzahl
git-svn-id: https://svn.r7r.at/sepm/trunk@327 240b89b8-79f3-4156-a269-5c34a98ffd57

+++ website/src/main/java/at/fakeroot/sepm/client/serialize/ObjectSearchServiceAsync.java
@@ -9,7 +9,7 @@ public interface ObjectSearchServiceAsync
 {
 	public void search(BoundingBox box, String what, AsyncCallback<SearchResult> callBack);
 	
-	public void getDetailHTML(int objId, AsyncCallback<ObjectDetails> callBack);
+	public void getDetailHTML(long objId, AsyncCallback<ObjectDetails> callBack);
 }
